refactor: correct folder name
