+++ packages/generator/src/fileCreators/idb-utils/filters/BoolFilter.ts
