+++ packages/generator/src/fileCreators/idb-utils/filters/BytesFilter.ts
