+++ packages/generator/src/fileCreators/idb-utils/filters/NumberFilter.ts
