+++ packages/generator/src/fileCreators/idb-utils/filters/StringFilter.ts
