+++ packages/generator/src/fileCreators/idb-utils/updateHandlers/BooleanHandler.ts
