+++ packages/generator/src/fileCreators/idb-utils/updateHandlers/BytesHandler.ts
