+++ packages/generator/src/fileCreators/idb-utils/updateHandlers/IntHandler.ts
