+++ packages/generator/src/fileCreators/idb-utils/updateHandlers/StringHandler.ts
